MetaMaterials/old.src/requirements.txt
inquirer 
#logging 
coloredlogs 
argparse 
pyyaml 
termcolor 
colorama 
termtables 
rapidfuzz 
scandir 
ipython 
pyperclip 